[rename] 카드이미지 파일 위치 이동
