+++ week3/src/App.jsx
@@ -1,9 +1,8 @@
 import "./App.css";
 import { ThemeProvider } from "styled-components";
 import MainPage from "./pages/CardGamePage";
-// import Header from "./components/header";
-import GlobalStyle from "./GlobalStyle";
-import theme from "./theme";
+import GlobalStyle from "./styles/GlobalStyle";
+import theme from "./styles/theme";
 
 function App() {
   return (
