+++ week3/src/assets/images/photoCard/nowitzki.webp
