+++ week3/src/contants/images.js
@@ -1,42 +1,62 @@
 export const IMAGEDATA = [
-  { id: 1, name: "12", imgUrl: "/card/12.webp", description: "12" },
-  { id: 2, name: "24:26", imgUrl: "/card/2426.webp", description: "24:26" },
+  {
+    id: 1,
+    name: "12",
+    imgUrl: "/src/assets/images/photoCard/12.webp",
+    description: "12",
+  },
+  {
+    id: 2,
+    name: "24:26",
+    imgUrl: "/src/assets/images/photoCard/2426.webp",
+    description: "24:26",
+  },
   {
     id: 3,
     name: "lifeInColors",
-    imgUrl: "/card/lifeInColors.jpeg",
+    imgUrl: "/src/assets/images/photoCard/lifeInColors.jpeg",
     description: "lifeInColors",
   },
   {
     id: 4,
     name: "lifeslike",
-    imgUrl: "/card/lifeslike.jpeg",
+    imgUrl: "/src/assets/images/photoCard/lifeslike.jpeg",
     description: "lifeslike",
   },
   {
     id: 5,
     name: "nowitzki",
-    imgUrl: "/card/nowitzki.webp",
+    imgUrl: "/src/assets/images/photoCard/nowitzki.webp",
     description: "nowitzki",
   },
   {
     id: 6,
     name: "WeAreGoingTo",
-    imgUrl: "/card/WeAreGoingTo.jpeg",
+    imgUrl: "/src/assets/images/photoCard/WeAreGoingTo.jpeg",
     description: "WeAreGoingTo",
   },
-  { id: 7, name: "waves", imgUrl: "/card/waves.jpeg", description: "waves" },
-  { id: 8, name: "soWhat", imgUrl: "/card/soWhat.jpeg", description: "soWhat" },
+  {
+    id: 7,
+    name: "waves",
+    imgUrl: "/src/assets/images/photoCard/waves.jpeg",
+    description: "waves",
+  },
+  {
+    id: 8,
+    name: "soWhat",
+    imgUrl: "/src/assets/images/photoCard/soWhat.jpeg",
+    description: "soWhat",
+  },
   {
     id: 9,
     name: "daliVan",
-    imgUrl: "/card/daliVan.jpeg",
+    imgUrl: "/src/assets/images/photoCard/daliVan.jpeg",
     description: "daliVan",
   },
   {
     id: 10,
     name: "alwaysAwake",
-    imgUrl: "/card/always.jpeg",
+    imgUrl: "/src/assets/images/photoCard/always.jpeg",
     description: "alwaysAwake",
   },
 ];
